Revert "Alerting the player is a separate function now"
This reverts commit da688be23110f9fbaa2c38077277f644177e7454.

This commit completely removes the use of aggregator alertUser
StringBuilder, and as a result of the constructs here will only alert
people of transforms / discoveries that have a customPrefix. If no
custom prefix for a drop is defined, no notification will ever be sent.
Aggregation of notifications is straight broken by this commit as well.

It also eliminates the " nearby" custom messaging of generation vs.
drop, just straight up gone?

I'll follow up this revert with a new commit to do the same but safely.